refactor dir
