+++ src/components/model/resume/pdf/ShortText.tsx
@@ -1,8 +1,8 @@
 import { Text, View } from '@react-pdf/renderer';
 import React from 'react';
 
-import Cell from './Cell';
-import Label from './Label';
+import Cell from './atoms/Cell';
+import Label from './atoms/Label';
 
 type Props = {
   label: string;
